% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust1d_package.R
\docType{package}
\name{hclust1d-package}
\alias{hclust1d-package}
\title{hclust1d-package}
\section{All \code{hclust1d} Functions}{


Similar in use to \code{stats::hclust}. It consists of the following functions:

\code{\link{hclust1d}} - univariate agglomerative hierarchical clustering routine.

\code{\link{supported_methods}} - lists all currently supported linkage methods.

\code{\link{supported_dist.methods}} - lists all currently supported distance methods.

For more information see a friendly "Getting started" vignette:
}

\examples{
\dontrun{
vignette("getting-started", package="hclust1d")
}

}
