% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust1d.R, R/hclust1d_package.R
\docType{package}
\name{hclust1d}
\alias{hclust1d}
\title{Hierarchical CLUSTering for 1D}
\usage{
hclust1d(points, method = "single")
}
\arguments{
\item{points}{set of 1D points to be clustered}

\item{method}{linkage method}
}
\value{
An object with S3 class \code{"hclust"}, compatible with a regular \code{stats::hclust} output:
}
\description{
1 dimensional hierarchical clustering  with a few possible choices of a linkage function.
}
\details{
1 dimensional hierarchical clustering  with a few possible choices of a linkage function.
}
\examples{
dendrogram <- hclust1d(rnorm(100))
plot(dendrogram)

}
