% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_methods.R
\name{supported_methods}
\alias{supported_methods}
\title{Supported Linkage Methods}
\usage{
supported_methods()
}
\value{
A character vector with currently supported linkage methods.
}
\description{
Lists all choices of a linkage method currently supported in \code{hclust1d}.
}
\examples{

if ("median" \%in\% supported_methods()) {    # the condition under if evaluates to TRUE
   dendrogram <- hclust1d(rnorm(100), method = "median")
   plot(dendrogram)
} else {
   stop("Error: median linkage method not supported in hclust1d")
}

}
