% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_dist.methods.R
\name{supported_dist.methods}
\alias{supported_dist.methods}
\title{Supported Distance Methods}
\usage{
supported_dist.methods()
}
\value{
A character vector with currently supported distance methods.
}
\description{
Lists all choices of a distance method currently supported in \code{hclust1d} via \code{dist} call.
}
